Clarify lesson programs (Requires EC approval)
This is a clarification to the definition of Lesson Programs that was brought up during Trainer Leadership meeting 18 October 2021. In attendance include @masamiy and @sstevens2 from the EC. I know any changes to our bylaws require a majority vote from the EC. @sstevens2 could you please bring this to the EC for vote or propose a different place to make this clarification?